requirements.txt
PyMuPDF~=1.24.11
opencv-python~=4.10.0.84
Pillow~=11.0.0
numpy~=2.1.2
openpyxl~=3.1.5
pytesseract~=0.3.13
ttkthemes~=3.2.2
pathvalidate
pip~=23.3
wheel~=0.41.2
setuptools~=70.0.0
et-xmlfile~=1.1.0
packaging~=24.1
pdf2image~=1.17.0